+++ raft-registry/src/store/config.rs
@@ -6,6 +6,6 @@ pub struct NodeConfig {
     pub snapshot_path: String,
     pub instance_prefix: String,
     pub journal_path: String,
-    pub snapshot_per_events: u32,
+    pub snapshot_per_events: u64,
     pub management_code: Option<String>,
 }
